requirements.txt
streamlit==1.28.0
opencv-python==4.8.1.78
fer==22.5.1
pandas==2.1.1
numpy==1.26.0
plotly==5.17.0
tensorflow==2.14.0
mtcnn==0.1.1
Pillow==10.0.1
mysql-connector-python
python-dotenv